move avatar to design-system
